+++ js/listview.js
@@ -27,6 +27,11 @@ enyo.kind({
 	create: function() {
 		this.inherited(arguments);
 		this.realLength = 0;
+		this.scrollbarTop = 0;
+		this.scrollbarMaxTop = 0;
+		this.move_scrollbar = true;
+		this.scrollbar_session_value = 0;
+		this.scroll_count = 0;
 		this.favoriteActivityButton = null;
 		if (!window.sugarizerOS) {
 			this.activitiesChanged();
@@ -41,6 +46,16 @@ enyo.kind({
 				t.draw();
 			});
 		}
+	},
+
+	rendered:function() {
+
+		this.inherited(arguments);
+		this.scrollbar_session_value = preferences.getScrollValue();
+		if (!this.scrollbar_session_value) {
+			this.scrollbar_session_value = 0;
+		}
+		this.scrollTo(0, this.scrollbar_session_value);
 	},
 
 	localize: function() {
@@ -115,7 +130,7 @@ enyo.kind({
 		inEvent.item.$.activity.setPopupShow(enyo.bind(this, "showActivityPopup"));
 		inEvent.item.$.activity.setPopupHide(enyo.bind(this, "hideActivityPopup"));
 		inEvent.item.$.favorite.setIcon({directory: "icons", icon: "emblem-favorite-large.svg"});
-		inEvent.item.$.favorite.setColorized(activitiesList[inEvent.index].favorite);
+		inEvent.item.$.favorite.setColorized(activities.getFavoritesName().indexOf(activitiesList[inEvent.index].id)!=-1);
 		inEvent.item.$.name.setContent(activitiesList[inEvent.index].name);
 		inEvent.item.$.version.setContent(l10n.get("VersionNumber", {number:activitiesList[inEvent.index].version}));
 		inEvent.item.$.help.setIcon({directory: "icons", icon: "help-rev.svg"});
@@ -129,13 +144,42 @@ enyo.kind({
 
 	// Handle scroll to lazy display content
 	onscroll: function(inSender, inEvent) {
+
 		var scrollBounds = inEvent.scrollBounds;
+		if(scrollBounds != undefined) {
+			this.scrollbarTop = Math.ceil(scrollBounds.top);
+			this.scrollbarMaxTop = Math.ceil(scrollBounds.maxTop);
+			if(!this.move_scrollbar) {
+				var diff = this.scrollbarTop - this.scrollbarMaxTop;
+				if (diff > 0) {
+					preferences.setScrollValue(this.scrollbarTop - diff);
+				}
+				else {
+					preferences.setScrollValue(this.scrollbarTop);
+				}
+				preferences.save()
+			}
+		}
+
 		var currentCount = this.$.activityList.get("count");
 		if (app.getToolbar().getSearchText().length == 0 && scrollBounds && (scrollBounds.maxTop - scrollBounds.top) < constant.listScrollLimit && this.realLength > currentCount) {
 			var length = Math.min(currentCount + constant.listStepCount, this.activities.length);
 			humane.log(l10n.get("Loading"));
 			this.$.activityList.set("count", length, true);
 		}
+
+		if (this.scrollbarTop <= this.scrollbar_session_value && this.move_scrollbar) {
+			this.scrollTo(0, this.scrollbar_session_value);
+		}
+
+		if (this.scrollbarTop == this.scrollbar_session_value && this.move_scrollbar) {
+			this.scroll_count = this.scroll_count + 1;
+		}
+
+		if (this.scrollbarTop > this.scrollbar_session_value || this.scroll_count == 2) {
+			this.move_scrollbar = false;
+		}
+
 	},
 
 	// Switch favorite value for clicked line
@@ -146,7 +190,7 @@ enyo.kind({
 	switchFavorite: function(favorite, activity) {
 		stats.trace(constant.viewNames[app.getView()], 'switch_favorite', activity.id, null);
 		util.vibrate();
-		favorite.setColorized(preferences.switchFavoriteActivity(activity));
+		favorite.setColorized(activities.switchFavoriteActivity(activity));
 		favorite.container.render();
 		preferences.save();
 		preferences.saveToServer(myserver);